+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.0] - 2022-07-22
 
+### Features
+
+- Initial Commit
+- Create ``` Virtual Network (VNET)```
+
+
 [Unreleased]: https://github.com/patrickhayo/azr-tf-module-vnet/compare/1.0.0...HEAD
 
 [1.0.0]: https://github.com/patrickhayo/azr-tf-module-vnet/compare/baa147eba84f61cd7ea6b4b76f8a4da8616cdc2e...1.0.0